Set up different file and stages for alinging from chains
